Remove: globalStore, observer move to observer folder
