--- src/utils/globalStore/GlobalStore.js
@@ -1,14 +1,0 @@
-import Observer from "./Observer";
-
-/* 
-
-*/
-const globalStore = {
-  state: {},
-  observer: new Observer(),
-};
-
-globalStore.setState = function (key, value) {
-  this.state[key] = value;
-  this.observer.notify({ key, value });
-};
