+++ src/utils/myRedux/createStore.js
@@ -1,6 +1,7 @@
-import Observer from "../globalStore/Observer.js";
+import Observer from "../observer/Observer.js";
 
 export const createStore = (reducer) => {
+  const observer = Object.freeze(new Observer());
   const state = reducer();
   return { subscribe, dispatch, getState };
 };
